mis a jour du projet avec integration de Spring security et affectation de role d'acces + version1 sans password encrypt

+++ src/main/java/fr/eni/pizzaOnLine/dao/RoleRepository.java
@@ -5,8 +5,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import fr.eni.pizzaOnLine.entities.Role;
 
-public interface RoleRepository extends JpaRepository<Role, Integer> {
+public interface RoleRepository extends JpaRepository<Role, Long> {
 
+	
 	Role findByLibelle(String libelle);
 	
 }
